Code cleanup -- remove old and abandoned classes, refactor where needed
