+++ api/src/makemkv_headless_api/api/v1/__init__.py
@@ -7,7 +7,6 @@ from . import socket
 from . import state
 from . import tmdb
 from . import toc
-from . import queue
 
 router = APIRouter(prefix="/v1")
 router.include_router(config.router)
@@ -17,4 +16,3 @@ router.include_router(socket.router)
 router.include_router(state.router)
 router.include_router(tmdb.router)
 router.include_router(toc.router)
-router.include_router(queue.router)
