+++ api/src/makemkv_headless_api/tmdb/__init__.py
@@ -4,7 +4,6 @@ import logging
 from typing import Any
 import requests
 
-from makemkv_headless_api.json_serializable import JSONSerializable
 from makemkv_headless_api.config import CONFIG
 from makemkv_headless_api.models.tmdb import TMDBConfigurationModel, TMDBMovieSearchResultModel, TMDBShowSearchResultModel
 
@@ -13,31 +12,6 @@ logger = logging.getLogger(__name__)
 API_HOST_NAME = 'api.themoviedb.org'
 WEB_HOST_NAME = 'www.themoviedb.org'
 API_VERSION = '3'
-
-class Data(JSONSerializable):
-  def __init__(self, data):
-    self.data = data
-
-  def __getattr__(self, attr: str):
-    try:
-      return self.data[attr]
-    except KeyError as ex:
-      if attr == 'name':
-        return self.data['title']
-      else:
-        raise ex
-  
-  def json_encoder(self):
-    return self.data
-
-class SearchResult(Data):
-  def __init__(self, data, content_type):
-    super().__init__(data)
-    self.content_type = content_type
-    self.url = f'https://{WEB_HOST_NAME}/{content_type}/{self.id}'
-
-  def __str__(self):
-    return f'{self.id}: {self.name}, {self.url}'
 
 def search(content, query):
   content = content.lower()
@@ -56,7 +30,7 @@ def search(content, query):
     elif content == 'tv':
       return [TMDBShowSearchResultModel(**result) for result in response.json()['results']]
   else:
-    return Data(response_json)
+    return [] # No results
 
 def configuration():
   url = f'https://{API_HOST_NAME}/{API_VERSION}/configuration'
