--- api/test/test_interface.py
@@ -1,39 +1,0 @@
-
-#!/usr/bin/env python3
-
-from unittest import TestCase
-from unittest.mock import patch
-from makemkv_headless_api.interface.target import Target
-from makemkv_headless_api.message.base_message_event import BaseMessageEvent
-from makemkv_headless_api.message.message_event import MessageEvent
-from makemkv_headless_api.message.progress_message_event import ProgressMessageEvent
-
-class InterfaceTest(TestCase):
-  def test_BaseMessageEvent(self):
-    base_message = BaseMessageEvent(target='foo', raw='bar', arbitrary='baz')
-    self.assertEqual('foo', base_message.target)
-    self.assertEqual('bar', base_message.raw)
-    self.assertEqual('baz', base_message.arbitrary)
-    self.assertEqual('{"target": "foo", "raw": "bar", "arbitrary": "baz", "type": "BaseMessageEvent"}', base_message.to_json())
-
-    self.assertEqual(base_message.to_json(), BaseMessageEvent.from_json(base_message.to_json()).to_json())
-    self.assertEqual(BaseMessageEvent, type(BaseMessageEvent.from_json(base_message.to_json())))
-
-    self.assertEqual(MessageEvent, type(BaseMessageEvent.from_json(MessageEvent('foo', target='bar').to_json())))
-
-  def test_MessageEvent(self):
-    self.assertEqual('this is a message\n', MessageEvent('this is a message', target='foo').text)
-    self.assertEqual('this is a message\n', MessageEvent('this', 'is', 'a', 'message', target='foo').text)
-
-    with self.assertRaises(AssertionError):
-      MessageEvent(target='foo')
-
-  def test_ProgressMessageEvent(self):
-    self.assertEqual(10, ProgressMessageEvent(total=10, current=30).total)
-    self.assertEqual(30, ProgressMessageEvent(total=10, current=30).current)
-
-    with self.assertRaises(AssertionError):
-      ProgressMessageEvent(total=10, target='foo')
-
-    with self.assertRaises(AssertionError):
-      ProgressMessageEvent(current=10, target='foo')
